Move inline_string and basic_string template helpers out
of a template_helpers.hpp. The template_helpers.hpp file should
not have any dependencies on libpmemobj-cpp headers.
